string: make strncmp() "count" type consistent

+++ string.h
@@ -70,7 +70,7 @@ extern int strcmp(const char *str1, const char *str2);
 
 
 /* Compares at most the first n bytes of str1 and str2. */
-extern int strncmp(const char *str1, const char *str2, unsigned int n);
+extern int strncmp(const char *str1, const char *str2, size_t n);
 
 
 /* Compares string str1 to str2. The result is dependent on the LC_COLLATE setting of the location. */
